Move DIDProfileDropdown out of NFTMoveToProfileDialog
